+++ test/refresh/refresh.config.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const config = _.clone(require('../default.config'));
 
-config.features = { refreshToken: true };
+config.features = { alwaysIssueRefresh: true };
 
 module.exports = {
   config,